Refactor Board, Field, Error
