+++ src/sudoku/mod.rs
@@ -1,0 +1,7 @@
+mod sudoku_board;
+mod sudoku_error;
+mod sudoku_field;
+
+pub use sudoku_board::SudokuBoard;
+pub use sudoku_error::SudokuError;
+pub use sudoku_field::SudokuField;
