+++ src/terminal_renderers/terminal_renderer.rs
@@ -1,4 +1,4 @@
-use fabrik::{renderers::SudokuRenderer, sudoku_board::SudokuBoard};
+use fabrik::{renderers::SudokuRenderer, sudoku::SudokuBoard};
 
 pub struct TerminalRenderer {}
